demo-app/requirements.txt
# GuardianAI Demo Application - Python Dependencies
flask==3.0.0
flask-cors==4.0.0
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
